+++ src/main/resources/sql/init.sql
@@ -22,14 +22,14 @@ CREATE TABLE inventory
 CREATE TABLE item
 (
     id                   BIGSERIAL PRIMARY KEY,
-    serial_number        BIGINT           NOT NULL,
-    name                 VARCHAR(64)      NOT NULL,
-    inventory_number     VARCHAR(24)      NOT NULL,
-    stored_in            VARCHAR(64)      NOT NULL,
-    unit                 VARCHAR(12)      NOT NULL,
-    quantity             DOUBLE PRECISION NOT NULL,
-    price                DOUBLE PRECISION NOT NULL,
-    is_owned_by_employee BOOLEAN          NOT NULL,
-    created_at           TIMESTAMP        NOT NULL,
-    created_by           INTEGER          NOT NULL REFERENCES users
+    serial_number        BIGINT        NOT NULL,
+    name                 VARCHAR(64)   NOT NULL,
+    inventory_number     VARCHAR(24)   NOT NULL,
+    stored_in            VARCHAR(64)   NOT NULL,
+    unit                 VARCHAR(12)   NOT NULL,
+    quantity             NUMERIC(8, 3) NOT NULL,
+    price                NUMERIC(8, 3) NOT NULL,
+    is_owned_by_employee BOOLEAN       NOT NULL,
+    created_at           TIMESTAMP     NOT NULL,
+    created_by           INTEGER       NOT NULL REFERENCES users
 );
